requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
python-dotenv==0.10.3
